refactor: declare a controller interface
