+++ src/main/java/com/yourssu/Application.java
@@ -1,10 +1,11 @@
 package com.yourssu;
 
-import com.yourssu.controller.BoardGameControllerExample;
+import com.yourssu.controller.implement.BoardGameControllerExample;
+import com.yourssu.controller.Controller;
 
 public class Application {
     public static void main(String[] args) {
-        Runnable controller = new BoardGameControllerExample();
+        Controller controller = new BoardGameControllerExample();
         controller.run();
     }
 }
