--- src/main/java/com/yourssu/controller/BoardGameController.java
@@ -1,7 +1,0 @@
-package com.yourssu.controller;
-
-public class BoardGameController implements Runnable {
-    @Override
-    public void run() {
-    }
-}
