+++ src/main/java/com/yourssu/controller/Controller.java
@@ -1,0 +1,6 @@
+package com.yourssu.controller;
+
+@FunctionalInterface
+public interface Controller {
+    void run();
+}
