+++ src/main/java/com/yourssu/controller/implement/BoardGameController.java
@@ -1,0 +1,9 @@
+package com.yourssu.controller.implement;
+
+import com.yourssu.controller.Controller;
+
+public class BoardGameController implements Controller {
+    @Override
+    public void run() {
+    }
+}
